gnosis meta transaction parsing
