+++ app/ts/components/simulationExplaining/customExplainers/EnsEventExplainer.tsx
@@ -1,6 +1,6 @@
+import { EnsEvent } from '../../../types/EnrichedEthereumData.js'
 import { RpcNetwork } from '../../../types/rpc.js'
 import { RenameAddressCallBack } from '../../../types/user-interface-types.js'
-import { EnsEvent } from '../../../types/visualizer-types.js'
 import { dataStringWith0xStart } from '../../../utils/bigint.js'
 import { assertNever } from '../../../utils/typescript.js'
 import { SmallAddress } from '../../subcomponents/address.js'
